services/worker/app/requirements.txt
boto3==1.34.144
gitpython==3.1.43
requests==2.32.3
PyJWT==2.9.0
# Claude（例：anthropic公式SDK）。実プロジェクトの採用SDKに合わせて固定してください
anthropic==0.36.0